Engine - change RawEntity.archIndex to int

+++ Engine/src/ECS/RawEntity/RawEntityStore.cs
@@ -109,8 +109,8 @@ public sealed class RawEntityStore : EntityStore
     {
         ref var entity      = ref entities[id];
         var archetype       = archetypes[entity.archIndex];
-        var result          =  RemoveComponent<T>(id, ref archetype, ref entity.compIndex);
-        entity.archIndex    = (short)archetype.archIndex;
+        var result          = RemoveComponent<T>(id, ref archetype, ref entity.compIndex);
+        entity.archIndex    = archetype.archIndex;
         return result;
     }
 }
